+++ models/__init__.py
@@ -1,0 +1,4 @@
+# models/__init__.py
+# 导入模型，使其可以通过models包直接访问
+
+from models.anime_model import Anime
