+++ utils/__init__.py
@@ -1,0 +1,5 @@
+# utils/__init__.py
+# 导入工具函数，使其可以通过utils包直接访问
+
+from utils.network import fetch_data_with_retry
+from utils.text_processor import preprocess_name
